renames rest
