+++ api/specs/web-server/_projects_nodes.py
@@ -26,13 +26,13 @@ from models_library.groups import GroupID
 from models_library.projects import ProjectID
 from models_library.projects_nodes_io import NodeID
 from simcore_service_webserver._meta import API_VTAG
-from simcore_service_webserver.projects._controller._crud_handlers import (
-    ProjectPathParams,
-)
 from simcore_service_webserver.projects._controller._nodes_rest import (
     NodePathParams,
     _ProjectGroupAccess,
     _ProjectNodePreview,
+)
+from simcore_service_webserver.projects._controller._projects_rest import (
+    ProjectPathParams,
 )
 
 router = APIRouter(
