+++ api/specs/web-server/_projects_states.py
@@ -14,7 +14,7 @@ from pydantic import ValidationError
 from servicelib.aiohttp import status
 from simcore_service_webserver._meta import API_VTAG
 from simcore_service_webserver.director_v2.exceptions import DirectorServiceError
-from simcore_service_webserver.projects._controller._states_handlers import (
+from simcore_service_webserver.projects._controller._projects_states_rest import (
     ProjectPathParams,
     _OpenProjectQuery,
 )
