+++ services/web/server/src/simcore_service_webserver/projects/_controller/_projects_rest.py
@@ -56,7 +56,7 @@ from .._crud_handlers_models import (
 from .._permalink_service import update_or_pop_permalink_in_project
 from ..models import ProjectDict
 from ..utils import get_project_unavailable_services, project_uses_available_services
-from . import _crud_handlers_utils
+from . import _projects_rest_utils
 
 # When the user requests a project with a repo, the working copy might differ from
 # the repo project. A middleware in the meta module (if active) will resolve
@@ -166,11 +166,11 @@ async def list_projects(request: web.Request):
         order_by=OrderBy.model_construct(**query_params.order_by.model_dump()),
     )
 
-    projects = await _crud_handlers_utils.aggregate_data_to_projects_from_request(
+    projects = await _projects_rest_utils.aggregate_data_to_projects_from_request(
         request, projects
     )
 
-    return _crud_handlers_utils.create_page_response(
+    return _projects_rest_utils.create_page_response(
         projects=projects,
         request_url=request.url,
         total=total_number_of_projects,
@@ -206,11 +206,11 @@ async def list_projects_full_search(request: web.Request):
         order_by=OrderBy.model_construct(**query_params.order_by.model_dump()),
     )
 
-    projects = await _crud_handlers_utils.aggregate_data_to_projects_from_request(
+    projects = await _projects_rest_utils.aggregate_data_to_projects_from_request(
         request, projects
     )
 
-    return _crud_handlers_utils.create_page_response(
+    return _projects_rest_utils.create_page_response(
         projects=projects,
         request_url=request.url,
         total=total_number_of_projects,
